requirements.txt
asyncpg~=0.21.0
aiogram==2.9.2
aiohttp==3.6.2
aioredis==1.3.1
python-dotenv~=0.14.0
tortoise-orm==0.16.15
sqlalchemy
gino
config~=0.5.0.post0